refact/front/GreenButton.jsx: moves to src/components/buttons/GreenButton.jsx
